feat(urlFavorites): implement import/export functionality with drag-and-drop support and format selection
